+++ types/elevationMode.ts
@@ -1,0 +1,3 @@
+export type ElevationMode = "heightAboveGround" | "heightAboveSeaLevel";
+
+export default ElevationMode;
