+++ types/event.ts
@@ -1,0 +1,6 @@
+export interface WrldEvent<T = string> {
+  type: T;
+  target: any;
+}
+
+export type EventHandler<T extends WrldEvent = WrldEvent<any>> = (e: T) => void
