+++ types/themes.ts
@@ -1,0 +1,25 @@
+declare namespace themes {
+  enum season {
+    Spring = "Spring",
+    Summer = "Summer",
+    Autumn = "Autumn",
+    Winter = "Winter",
+  }
+
+  enum time {
+    Dawn = "Dawn",
+    Day = "Day",
+    Dusk = "Dusk",
+    Night = "Night",
+  }
+
+  enum weather {
+    Clear = "Default",
+    Overcast = "Overcast",
+    Foggy = "Foggy",
+    Rainy = "Rainy",
+    Snowy = "Snowy",
+  }
+}
+
+export default themes;
