+++ types/vector.ts
@@ -1,0 +1,12 @@
+export type Vector3 = [number, number, number] | {
+  x: number;
+  y: number;
+  z: number;
+};
+
+export type Vector4 = [number, number, number, number] | {
+  x: number;
+  y: number;
+  z: number;
+  w: number;
+};
